Move VariantSelector to separate file
