+++ src/scenes/ProgressReports/EditForm/Steps/ProgressStep/ProgressStep.tsx
@@ -1,11 +1,5 @@
 import { useMutation } from '@apollo/client';
-import {
-  Box,
-  Grid,
-  ToggleButton,
-  ToggleButtonGroup,
-  Typography,
-} from '@mui/material';
+import { Box, Grid, Typography } from '@mui/material';
 import { groupBy, isEmpty, sortBy } from 'lodash';
 import { useMemo, useState } from 'react';
 import { VariantFragment as Variant } from '~/common/fragments';
@@ -19,36 +13,12 @@ import { ProgressSummaryCard } from '../../../../ProgressReports/Detail/Progress
 import { UpdatePeriodicReportDialog } from '../../../../Projects/Reports/UpdatePeriodicReportDialog';
 import { ProductTable } from '../../../Detail/ProductTable';
 import { ProgressReportCard } from '../../../Detail/ProgressReportCard';
-import { colorPalette } from '../../colorPalette';
 import { useProgressReportContext } from '../../ProgressReportContext';
 import {
   ProgressReportEditFragment,
   ProgressReportProgressFragment,
 } from '../../ProgressReportEdit.graphql';
-import { RoleIcon } from '../../RoleIcon';
-
-const transparentBgSx = {
-  backgroundColor: 'transparent',
-};
-
-const iconSx = {
-  height: 36,
-  width: 36,
-  mr: 0,
-};
-
-const ToggleButtonSx = (role: string) => ({
-  p: 0.25,
-  pr: 1,
-  mr: 0,
-  borderRadius: 0.6,
-  '&.Mui-selected': {
-    backgroundColor: colorPalette.stepperCard.iconBackground[role],
-  },
-  '&.Mui-selected:hover': {
-    backgroundColor: colorPalette.stepperCard.iconBackground[role],
-  },
-});
+import { VariantSelector } from './VariantSelector';
 
 export const ProgressStep = () => {
   const { report } = useProgressReportContext();
@@ -165,10 +135,10 @@ const EditableProductTable = ({
       pagination
       header={() => (
         <div style={{ display: 'flex', justifyContent: 'space-between' }}>
-          <VariantsToggle
-            variant={variant}
+          <VariantSelector
             variants={variants}
-            setVariant={setVariant}
+            value={variant}
+            onChange={setVariant}
           />
         </div>
       )}
@@ -191,52 +161,5 @@ const EditableProductTable = ({
         });
       }}
     />
-  );
-};
-
-const VariantsToggle = ({
-  variants,
-  variant,
-  setVariant,
-}: {
-  variant: Variant;
-  setVariant: (variant: Variant) => void;
-  variants: Variant[];
-}) => {
-  const isSelected = (v: Variant) => v.key === variant.key;
-
-  if (variants.length < 2) {
-    return null;
-  }
-
-  return (
-    <ToggleButtonGroup
-      value={variant.key}
-      exclusive
-      onChange={(_e, value) => {
-        if (value !== null) {
-          setVariant(variants.find((v) => v.key === value) ?? variants[0]!);
-        }
-      }}
-      color="secondary"
-      sx={{ mt: 1, ml: 1 }}
-    >
-      {variants.map((v) => (
-        <ToggleButton
-          key={v.key}
-          value={v.key}
-          sx={[
-            ToggleButtonSx(v.responsibleRole!),
-            isSelected(v) && transparentBgSx,
-          ]}
-        >
-          <RoleIcon
-            variantRole={v.responsibleRole}
-            sx={[iconSx, !isSelected(v) && transparentBgSx]}
-          />
-          {v.label}
-        </ToggleButton>
-      ))}
-    </ToggleButtonGroup>
   );
 };
